perf: 表达式解析器增加对流水线变量处理的兼容 #10609
